Sorted more utilities
