+++ SoulsFormats/Binder/BND3/BND3.cs
@@ -50,7 +50,7 @@ namespace SoulsFormats
         public BND3()
         {
             Files = new List<BinderFile>();
-            Version = SFUtil.DateToBinderTimestamp(DateTime.Now);
+            Version = Binder.DateToBinderTimestamp(DateTime.Now);
             Format = Binder.Format.IDs | Binder.Format.Names1 | Binder.Format.Names2 | Binder.Format.Compression;
         }
