+++ SoulsFormats/Binder/BND4/BND4.cs
@@ -59,7 +59,7 @@ namespace SoulsFormats
         public BND4()
         {
             Files = new List<BinderFile>();
-            Version = SFUtil.DateToBinderTimestamp(DateTime.Now);
+            Version = Binder.DateToBinderTimestamp(DateTime.Now);
             Format = Binder.Format.IDs | Binder.Format.Names1 | Binder.Format.Names2 | Binder.Format.Compression;
             Unicode = true;
             Extended = 4;
