+++ SoulsFormats/Utilities/Cryptography/HashHelper.cs
@@ -1,0 +1,42 @@
+﻿using System.Linq;
+
+namespace SoulsFormats.Utilities.Cryptography
+{
+    /// <summary>
+    /// Helper methods for hashing.
+    /// </summary>
+    public static class HashHelper
+    {
+        /// <summary>
+        /// FromSoft's basic filename hashing algorithm, used in some BND and BXF formats.
+        /// </summary>
+        public static uint FromPathHash(string text)
+        {
+            string hashable = text.ToLowerInvariant().Replace('\\', '/');
+            if (!hashable.StartsWith("/"))
+                hashable = '/' + hashable;
+            return hashable.Aggregate(0u, (i, c) => i * 37u + c);
+        }
+
+        /// <summary>
+        /// Determines whether a number is prime or not.
+        /// </summary>
+        public static bool IsPrime(uint candidate)
+        {
+            if (candidate < 2)
+                return false;
+            if (candidate == 2)
+                return true;
+            if (candidate % 2 == 0)
+                return false;
+
+            for (int i = 3; i * i <= candidate; i += 2)
+            {
+                if (candidate % i == 0)
+                    return false;
+            }
+
+            return true;
+        }
+    }
+}
