+++ SoulsFormats/Utilities/Formats/ISoulsFile.cs
