+++ SoulsFormats/Utilities/Formats/SoulsFile.cs
