+++ SoulsFormats/Utilities/SFUtil.cs
@@ -33,75 +33,6 @@ namespace SoulsFormats
         }
 
         /// <summary>
-        /// FromSoft's basic filename hashing algorithm, used in some BND and BXF formats.
-        /// </summary>
-        public static uint FromPathHash(string text)
-        {
-            string hashable = text.ToLowerInvariant().Replace('\\', '/');
-            if (!hashable.StartsWith("/"))
-                hashable = '/' + hashable;
-            return hashable.Aggregate(0u, (i, c) => i * 37u + c);
-        }
-
-        /// <summary>
-        /// Determines whether a number is prime or not.
-        /// </summary>
-        public static bool IsPrime(uint candidate)
-        {
-            if (candidate < 2)
-                return false;
-            if (candidate == 2)
-                return true;
-            if (candidate % 2 == 0)
-                return false;
-
-            for (int i = 3; i * i <= candidate; i += 2)
-            {
-                if (candidate % i == 0)
-                    return false;
-            }
-
-            return true;
-        }
-
-        private static readonly Regex timestampRx = new Regex(@"(\d\d)(\w)(\d+)(\w)(\d+)");
-
-        /// <summary>
-        /// Converts a BND/BXF timestamp string to a DateTime object.
-        /// </summary>
-        public static DateTime BinderTimestampToDate(string timestamp)
-        {
-            Match match = timestampRx.Match(timestamp);
-            if (!match.Success)
-                throw new InvalidDataException("Unrecognized timestamp format.");
-
-            int year = Int32.Parse(match.Groups[1].Value) + 2000;
-            int month = match.Groups[2].Value[0] - 'A';
-            int day = Int32.Parse(match.Groups[3].Value);
-            int hour = match.Groups[4].Value[0] - 'A';
-            int minute = Int32.Parse(match.Groups[5].Value);
-
-            return new DateTime(year, month, day, hour, minute, 0);
-        }
-
-        /// <summary>
-        /// Converts a DateTime object to a BND/BXF timestamp string.
-        /// </summary>
-        public static string DateToBinderTimestamp(DateTime dateTime)
-        {
-            int year = dateTime.Year - 2000;
-            if (year < 0 || year > 99)
-                throw new InvalidDataException("BND timestamp year must be between 2000 and 2099 inclusive.");
-
-            char month = (char)(dateTime.Month + 'A');
-            int day = dateTime.Day;
-            char hour = (char)(dateTime.Hour + 'A');
-            int minute = dateTime.Minute;
-
-            return $"{year:D2}{month}{day}{hour}{minute}".PadRight(8, '\0');
-        }
-
-        /// <summary>
         /// Compresses data and writes it to a BinaryWriterEx with Zlib wrapper.
         /// </summary>
         public static int WriteZlib(BinaryWriterEx bw, byte formatByte, byte[] input)
